refactored product obj - change parameter datatypes

+++ backend/superprice/src/main/java/com/example/superprice/model/Product.java
@@ -1,4 +1,4 @@
 package com.example.superprice.model;
 
-public record Product(Long id, String name, String description, String store, String imageURL, Long price, Long quantity) {
+public record Product(Long id, String name, String description, String store, String imageURL, double price, int quantity) {
 }
